refactor: Split logic from log index page
